Fix ACL Scale tests
After the introduction of a commit (1d4dc99) that allowed to
create traffic items with increment value fields the behavior
for specifying src/dst ports has changed. The port should only be
specified as a single value string, or as a dict with a type
(increment/decrement/random/list) and type parameters.

Signed-off-by: Serhiy Boiko <serhiy.boiko@plvision.eu>