Update cleanup_ip_addrs to support IPv6
IPv6 requires a link local address (fe80::/64) for the NDP service
to work properly. When `ip address flush` is called the link local
address is also flushed. The only way to restore it is to set the
operstate of the port down then up. As a side effect this will also
clear any port neighbors.
Since NDP will now work properly the IPv6 tests no longer require
sending NS/ping to/from the DUT (which was required to refresh
DUT neighbors).

Signed-off-by: Serhiy Boiko <serhiy.boiko@plvision.eu>