Fix QoS default prio test
Add traffic from ports 2, 3, 4 so the fdb entries do not age.

At the end of the test the fdb entries for ports 2, 3, 4
are aged since there is no traffic from TG to them. Since
there is no fdb entries then any traffic from port 1 will
be flooded instead of forwarded. And since the traffic is
flooded the qdisc counters are counting extra packets which
causes the test to fail.

Signed-off-by: Serhiy Boiko <serhiy.boiko@plvision.eu>